services/price-forecasting-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
psycopg2-binary==2.9.9
pydantic==2.5.2
python-dotenv==1.0.0
httpx==0.25.2
numpy==1.24.3
pandas==2.0.3
prophet==1.1.5
xgboost==2.0.3
scipy==1.11.4
cmdstanpy==1.1.0

